requirements.txt
flask==3.0.0
openai
httpx==0.27.2
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
